+++ spec/public_api_compatibility_spec.rb
@@ -52,6 +52,7 @@ RSpec.describe "Public API compatibility" do
       UiConfig
       UiConfigBuilder
       GraphAdapter
+      NodePresenter
       PathTree
       PathTreeBuilder
       ReverseTree